refactor(configuration): rename resolve module to config_value and update imports
Renamed `src/configuration/resolve.py` to `src/configuration/config_value.py` to better reflect its purpose. Updated all imports across the codebase to use the new module path, ensuring consistency and clarity in the configuration module structure.
