+++ ingesters/sead/policies.py
@@ -8,7 +8,7 @@ import pandas as pd
 from loguru import logger
 from pandas._typing import Dtype
 
-from src.configuration.resolve import ConfigValue
+from src.configuration import ConfigValue
 
 from .metadata import SchemaService, SeadSchema, Table
 from .utility import Registry, pascal_to_snake_case, snake_to_pascal_case
